Backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api 
flask[async]
flask
flask-cors
unicorn 
huggingface_hub